# services/gateway/docs/swagger.yaml
components:
  schemas:
    dto.CreateEventRequest:
      properties:
        artist:
          items:
            type: string
          type: array
          uniqueItems: false
        description:
          type: string
        eventDate:
          type: string
        images:
          items:
            type: string
          type: array
          uniqueItems: false
        locationId:
          type: string
        name:
          type: string
        thumbnail:
          type: string
      required:
      - artist
      - description
      - eventDate
      - images
      - locationId
      - name
      - thumbnail
      type: object
    dto.CreateEventResponse:
      properties:
        id:
          type: string
      required:
      - id
      type: object
    dto.CreateEventZoneRequest:
      properties:
        color:
          type: string
        description:
          type: string
        eventID:
          type: string
        locationId:
          type: string
        name:
          type: string
        price:
          type: number
        zoneNumber:
          type: integer
      required:
      - color
      - description
      - eventID
      - locationId
      - name
      - price
      - zoneNumber
      type: object
    dto.CreateEventZoneResponse:
      properties:
        id:
          type: string
      required:
      - id
      type: object
...
    dto.EventListResponse:
      properties:
        list:
          items:
            $ref: '#/components/schemas/dto.EventResponse'
          type: array
          uniqueItems: false
      required:
      - list
      type: object
    dto.EventResponse:
      properties:
        artist:
          items:
            type: string
          type: array
          uniqueItems: false
        createdAt:
          type: string
        description:
          type: string
        eventDate:
          type: string
        id:
          type: string
        images:
          items:
            type: string
          type: array
          uniqueItems: false
        locationId:
          type: string
        name:
          type: string
        thumbnail:
          type: string
        updatedAt:
          type: string
      required:
      - artist
      - createdAt
      - description
      - eventDate
      - id
      - images
      - locationId
      - name
      - thumbnail
      - updatedAt
      type: object
    dto.EventZoneListResponse:
      properties:
        list:
          items:
            $ref: '#/components/schemas/dto.EventZoneResponse'
          type: array
          uniqueItems: false
      required:
      - list
      type: object
    dto.EventZoneResponse:
      properties:
        color:
          type: string
        description:
          type: string
        eventId:
          type: string
        id:
          type: string
        isSoldOut:
          type: boolean
        locationId:
          type: string
        name:
          type: string
        price:
          type: number
        zoneNumber:
          type: integer
      required:
      - color
      - description
      - eventId
      - id
      - isSoldOut
      - locationId
      - name
      - price
      - zoneNumber
      type: object
    dto.HttpError:
      properties:
        error:
          type: string
      required:
      - error
      type: object
    dto.HttpResponse-dto_CreateEventResponse:
      properties:
        result:
          $ref: '#/components/schemas/dto.CreateEventResponse'
      required:
      - result
      type: object
    dto.HttpResponse-dto_CreateEventZoneResponse:
      properties:
        result:
          $ref: '#/components/schemas/dto.CreateEventZoneResponse'
      required:
      - result
      type: object
...
    dto.HttpResponse-dto_EventListResponse:
      properties:
        result:
          $ref: '#/components/schemas/dto.EventListResponse'
      required:
      - result
      type: object
    dto.HttpResponse-dto_EventResponse:
      properties:
        result:
          $ref: '#/components/schemas/dto.EventResponse'
      required:
      - result
      type: object
    dto.HttpResponse-dto_EventZoneListResponse:
      properties:
        result:
          $ref: '#/components/schemas/dto.EventZoneListResponse'
      required:
      - result
      type: object
...
    dto.HttpResponse-dto_LoginResponse:
      properties:
        result:
          $ref: '#/components/schemas/dto.LoginResponse'
      required:
      - result
      type: object
    dto.HttpResponse-dto_RefreshTokenResponse:
      properties:
        result:
          $ref: '#/components/schemas/dto.RefreshTokenResponse'
      required:
      - result
      type: object
    dto.HttpResponse-dto_UpdateEventResponse:
      properties:
        result:
          $ref: '#/components/schemas/dto.UpdateEventResponse'
      required:
      - result
      type: object
    dto.HttpResponse-dto_UpdateEventZoneResponse:
      properties:
        result:
          $ref: '#/components/schemas/dto.UpdateEventZoneResponse'
      required:
      - result
      type: object
...
    dto.LoginRequest:
      properties:
        idToken:
          type: string
        provider:
          type: string
      type: object
    dto.LoginResponse:
      properties:
        accessToken:
          type: string
        exp:
          type: integer
        isNewUser:
          type: boolean
        refreshToken:
          type: string
        user:
          $ref: '#/components/schemas/dto.UserResponse'
      required:
      - accessToken
      - exp
      - isNewUser
      - refreshToken
      - user
      type: object
    dto.RefreshTokenRequest:
      properties:
        refreshToken:
          type: string
      required:
      - refreshToken
      type: object
    dto.RefreshTokenResponse:
      properties:
        accessToken:
          type: string
        exp:
          type: integer
        refreshToken:
          type: string
      required:
      - accessToken
      - exp
      - refreshToken
      type: object
    dto.UpdateEventRequest:
      properties:
        artist:
          items:
            type: string
          type: array
          uniqueItems: false
        description:
          type: string
        eventDate:
          type: string
        images:
          items:
            type: string
          type: array
          uniqueItems: false
        locationId:
          type: string
        name:
          type: string
        thumbnail:
          type: string
      type: object
    dto.UpdateEventResponse:
      properties:
        id:
          type: string
      required:
      - id
      type: object
    dto.UpdateEventZoneRequest:
      properties:
        color:
          type: string
        description:
          type: string
        eventId:
          type: string
        isSoldOut:
          type: boolean
        locationId:
          type: string
        name:
          type: string
        price:
          type: number
        zoneNumber:
          type: integer
      type: object
    dto.UpdateEventZoneResponse:
      properties:
        id:
          type: string
      required:
      - id
      type: object
...
    dto.UserResponse:
      properties:
        birthdate:
          type: string
        createdAt:
          type: string
        deletedAt:
          type: string
        email:
          type: string
        firstname:
          type: string
        id:
          type: string
        lastname:
          type: string
        phone:
          type: string
        profileImage:
          type: string
        provider:
          type: string
        role:
          type: string
        updatedAt:
          type: string
      required:
      - birthdate
      - createdAt
      - email
      - firstname
      - id
      - lastname
      - phone
      - profileImage
      - provider
      - role
      - updatedAt
      type: object
...
  securitySchemes:
    ApiKeyAuth:
      in: header
      name: Authorization
      type: apiKey
...
info:
  description: A Concert Gateway API Documentation
  title: A Concert Gateway
  version: 1.0.0
openapi: 3.1.0
paths:
  /v1/auth/login:
    post:
      description: Login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/dto.LoginRequest'
        description: request request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpResponse-dto_LoginResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpError'
          description: Internal Server Error
      summary: Login
      tags:
      - auth
  /v1/auth/logout:
    post:
      description: Logout
      responses:
        "204":
          description: No Content
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpError'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpError'
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Logout
      tags:
      - auth
  /v1/auth/refresh:
    post:
      description: Refresh Token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/dto.RefreshTokenRequest'
        description: request request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpResponse-dto_RefreshTokenResponse'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpError'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpError'
          description: Internal Server Error
      summary: Refresh Token
      tags:
      - auth
  /v1/events:
    get:
      description: List Events
      parameters:
      - description: query
        in: query
        name: query
        schema:
          type: integer
      - description: sortBy
        in: query
        name: sortBy
        schema:
          type: integer
      - description: order
        in: query
        name: order
        schema:
          type: string
      - description: page
        in: query
        name: page
        schema:
          type: string
      - description: limit
        in: query
        name: limit
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpResponse-dto_EventListResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpError'
          description: Internal Server Error
      summary: List Events
      tags:
      - events
    post:
      description: Create Event
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/dto.CreateEventRequest'
        description: Create event request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpResponse-dto_CreateEventResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpError'
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Create Event
      tags:
      - events
  /v1/events/{id}:
    delete:
      description: Delete Event
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpError'
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Delete Event
      tags:
      - events
    get:
      description: Get Event
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpResponse-dto_EventResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpError'
          description: Internal Server Error
      summary: Get Event
      tags:
      - events
    put:
      description: Update Event
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/dto.UpdateEventRequest'
        description: Update event request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpResponse-dto_UpdateEventResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpError'
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Update Event
      tags:
      - events
  /v1/events/{id}/event-zones:
    get:
      description: Get Event Zones by Event ID
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpResponse-dto_EventZoneListResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpError'
          description: Internal Server Error
      summary: Get Event Zones by Event ID
      tags:
      - event-zones
    post:
      description: Create Event Zone
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/dto.CreateEventZoneRequest'
        description: Create event zone request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpResponse-dto_CreateEventZoneResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpError'
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Create Event Zone
      tags:
      - event-zones
  /v1/events/event-zones/{id}:
    delete:
      description: Delete Event Zone
      parameters:
      - description: Event Zone ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpError'
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Delete Event Zone
      tags:
      - event-zones
    put:
      description: Update Event Zone
      parameters:
      - description: Event Zone ID
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/dto.UpdateEventZoneRequest'
        description: Update event zone request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpResponse-dto_UpdateEventZoneResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.HttpError'
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Update Event Zone
      tags:
      - event-zones
...